code optimzing

+++ tasks/api/my_project_tasks/views.py
@@ -1,12 +1,13 @@
-from tasks.api.serializers import (
-    TaskSerializer, LessFieldsTaskSerializer, CommentSerializer, TaskListSerializer, TaskTrashedSerializer)
-from common.permissions_scopes import TaskPermissions, ProjectCommentPermissions, TaskCommentPermissions
-from tasks.actions import (
-    delete_dependencies, excluded_users, progress, tasksOfProject, tasksResponse, create, update)
 from common.comments import listComments, createComments, updateComments, broadcastDeleteComment
-from common.actions import (
-    delete, allItems, filterRecords, addAttachment, deleteAttachments, getAttachments)
+from common.actions import (delete, addAttachment,
+                            deleteAttachments, getAttachments, filterRecords, un_authorized)
+from tasks.actions import (
+    delete_dependencies, excluded_users, progress, tasksOfProject, create, update)
+from tasks.api.serializers import (
+    TaskSerializer, CommentSerializer, TaskListSerializer, TaskTrashedSerializer)
+from rest_framework.permissions import IsAuthenticated
 from common.custom import CustomPageNumberPagination
+from common.permissions import checkProjectScope
 from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework import viewsets, status
@@ -14,12 +15,12 @@ from common.Repository import Repository
 from tasks.models import Task, Comment
 
 
-class TaskViewSet(Repository):
+class MyTaskViewSet(Repository):
     model = Task
     queryset = Task.objects.filter(
         deleted_at__isnull=True).order_by("-created_at")
     serializer_class = TaskSerializer
-    permission_classes = (TaskPermissions,)
+    permission_classes = (IsAuthenticated,)
     serializer_action_classes = {
         "retrieve": TaskListSerializer,
         "update": TaskListSerializer,
@@ -31,49 +32,63 @@ class TaskViewSet(Repository):
 
     def list(self, request):
         queryset = self.get_queryset()
-        queryset = filterRecords(queryset, request, table=Task)
         if request.GET.get("project_id"):
-            return tasksOfProject(self, request, queryset)
-        if request.GET.get("items_per_page") == "-1":
-            return allItems(LessFieldsTaskSerializer, queryset)
-        if request.GET.get("items_per_page") == "-2":
-            return allItems(self.get_serializer, queryset)
-
-        page = self.paginate_queryset(queryset)
-        serializer = self.get_serializer(
-            page, many=True, context={"request": request})
-        return tasksResponse(self, serializer)
+            if checkProjectScope(request.user, None, "project_tasks_v", project_id=request.GET.get("project_id")):
+                queryset = filterRecords(queryset, request, table=Task)
+                return tasksOfProject(self, request, queryset)
+            else:
+                return un_authorized()
+        return un_authorized()
 
     def retrieve(self, request, pk=None):
         task = self.get_object()
-        serializer = self.get_serializer(
-            task, context={"request": request})
-        data = serializer.data
-        data = getAttachments(request, data, task.id, "task_attachments_v")
-        return Response(data)
+        if checkProjectScope(request.user, task.project, "project_tasks_v"):
+            serializer = self.get_serializer(
+                task, context={"request": request})
+            data = serializer.data
+            data = getAttachments(request, data, task.id,
+                                  "task_attachments_v", project=task.project)
+            return Response(data)
+        else:
+            return un_authorized()
 
     def create(self, request):
-        return create(request)
+        if checkProjectScope(request.user, None, "project_tasks_c", project_id=request.data.get("project")):
+            return create(request)
+        else:
+            return un_authorized()
 
     def update(self, request, pk=None):
         task = self.get_object()
-        return update(self, request, task)
+        if checkProjectScope(request.user, task.project, "project_tasks_u"):
+            return update(self, request, task)
+        else:
+            return un_authorized()
 
     @action(detail=True, methods=["get"])
     def excluded_users(self, request, pk=None):
         task = self.get_object()
-        return excluded_users(self, task)
+        if checkProjectScope(request.user, task.project, "project_tasks_v"):
+            return excluded_users(self, task)
+        else:
+            return un_authorized()
 
     @action(detail=True, methods=["delete"])
     def delete(self, request, pk=None):
         task = self.get_object()
-        return delete_dependencies(request, task)
+        if checkProjectScope(request.user, task.project, "project_tasks_u"):
+            return delete_dependencies(request, task)
+        else:
+            return un_authorized()
 
     @action(detail=True, methods=["post"])
     def add_attachments(self, request, pk=None):
         try:
             task = self.get_object()
-            return addAttachment(request, task)
+            if checkProjectScope(request.user, task.project, "task_attachments_c"):
+                return addAttachment(request, task)
+            else:
+                return un_authorized()
         except:
             return Response(
                 {"message": "something went wrong"}, status=status.HTTP_400_BAD_REQUEST
@@ -81,13 +96,20 @@ class TaskViewSet(Repository):
 
     @action(detail=True, methods=["delete"])
     def delete_attachments(self, request, pk=None):
-        return deleteAttachments(self, request)
+        task = self.get_object()
+        if checkProjectScope(request.user, task.project, "task_attachments_d"):
+            return deleteAttachments(self, request)
+        else:
+            return un_authorized()
 
     @action(detail=True, methods=["put"])
     def progress(self, request, pk=None):
         try:
             task = self.get_object()
-            return progress(request, task)
+            if checkProjectScope(request.user, task.project, "project_tasks_u"):
+                return progress(request, task)
+            else:
+                return un_authorized()
         except:
             return Response({'error': "Something went wrong"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
@@ -96,7 +118,7 @@ class ProjectCommentViewSet(viewsets.ModelViewSet):
     queryset = Comment.objects.all()
     serializer_class = CommentSerializer
     pagination_class = CustomPageNumberPagination
-    permission_classes = (ProjectCommentPermissions, )
+    permission_classes = (IsAuthenticated, )
 
     def list(self, request):
         return listComments(self, request)
@@ -117,7 +139,7 @@ class TaskCommentViewSet(viewsets.ModelViewSet):
     queryset = Comment.objects.all()
     serializer_class = CommentSerializer
     pagination_class = CustomPageNumberPagination
-    permission_classes = (TaskCommentPermissions,)
+    permission_classes = (IsAuthenticated,)
 
     def list(self, request):
         return listComments(self, request)
